Reviewed-on: https://codeberg.org/pr0game/pr0game/pulls/490

+++ includes/classes/class.PlanetRessUpdate.php
@@ -904,7 +904,7 @@ class ResourceUpdate
         Database::get()->update($sql, $params);
     }
 
-    public function addResources(int $planetId, array $resources, array &$planet = null, int $factor = 1)
+    public function addResources(int $planetId, array $resources, array &$planet = null, float $factor = 1)
     {
         global $resource;
         $params = [':planetId' => $planetId];
